auth/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
psycopg2-binary
passlib
python-jose
pydantic
bcrypt==4.0.1
twilio
python-dotenv
asyncpg